requirements.txt
requests
pandas
numpy
RPi.GPIO 
paho-mqtt